Fix create account
